--- config/db.php
@@ -1,4 +1,0 @@
-<?php
-require_once 'config/config.php';
-$db = new PDO(DB_DSN, DB_USER, DB_PASS, DB_OPTIONS);
-
