+++ public/index.php
@@ -4,7 +4,7 @@ use Blog\Router;
 use Tracy\Debugger;
 
 require_once ('../vendor/autoload.php');
-require_once ('../config/functions.php');
+require_once ('../src/Controllers/functions.php');
 
 $router = new Router();
