+++ src/Controllers/functions.php
@@ -1,0 +1,14 @@
+<?php
+
+function str_secur($string){
+    return trim(htmlspecialchars($string));
+}
+
+
+function debug($var){
+    echo '<pre>';
+    var_dump($var);
+    echo '</pre>';
+    die();
+
+}
