+++ src/Models/ArticlesModel.php
@@ -2,7 +2,6 @@
 namespace Blog\Models;
 
 class ArticlesModel extends MainModel{
-public $id;
 
 public function __construct()
 {
