+++ src/Models/HomeModel.php
@@ -3,7 +3,6 @@ namespace Blog\Models;
 
 class HomeModel extends MainModel{
 
-public $id;
 
 public function __construct()
 {
@@ -14,11 +13,4 @@ public function __construct()
     $this->getConnection();
 }
 
-public function getOne($id){
-    $req = "SELECT * FROM ". $this->table ." WHERE id='". $id ."'";
-    // $req = "SELECT * FROM ". $this->table ." WHERE id='10'";
-    $query = $this->_connexion->prepare($req);
-    $query->execute();
-    return $query->fetch();
-}
 }
