+++ src/Views/Article_edit.twig
@@ -11,10 +11,10 @@
    <p><input type="hidden"  name="id"         value="{{ article.id }}"></p>
    <p><input type="text"    name="title"      value="{{ article.title }}" id="defaultContactFormName" class="form-control mb-4" placeholder="Titre" required></p>
    <p><input type="text"    name="slug"       value="{{ article.slug }}" id="defaultContactFormName" class="form-control mb-4" placeholder="Slug" required></p>
-   <p><textarea             name="content"    value "{{ article.content }}" id="exampleFormControlTextarea2" class="form-control rounded-0"  rows="3" placeholder="Description" required></textarea></p> 
+   <p><input type="text"    name="content"       value="{{ article.content }}" id="defaultContactFormName" class="form-control mb-4" placeholder="content" required></p>
+   {# <p><textarea             name="content"    value="{{ article.content }}" id="exampleFormControlTextarea2" class="form-control rounded-0"  rows="3" placeholder="Description" required></textarea></p>  #}
    <p><input type="text"    name="url_images" value="{{ article.url_images }}" id="defaultContactFormName" class="form-control mb-4" placeholder="copier l'url de l'image" required></p>
-   <p><input type="submit"  value="editer"></p>
+   <p><input type="submit"  value="modifier"></p>
 
 </form>
 {% endblock content %}
-
