+++ src/Views/Template/layout.twig
@@ -32,6 +32,9 @@
 					<div class="col-8"> {% block content %}{% endblock %}</div>
 					<div class="col-4"> {% block rss %}{% endblock %}</div>
 				</div>
+				<div class="row">
+					<div class="col-6"> {% block form %}{% endblock %}</div>
+				</div>
 			</div>
 		</section>
 {#
